refactor(parser): separate search parsers into a single file
